Replaced last-minute hacks with maintainable code (#217)
* Fixed bug with switching models when there was insufficient data

* Abstracted rollout analyser

* Made parameter serialisation reversible

* Fixed bugs in parameters deserialisation

+++ pvg/trainers/ei_pure_text.py
@@ -6,17 +6,17 @@ from time import sleep
 import pickle
 
 import numpy as np
-from numpy import ma
 
 from jaxtyping import Bool
 
 from pvg.trainers.rl_pure_text_base import PureTextRlTrainer
 from pvg.trainers.registry import register_trainer
 from pvg.parameters import TrainerType
+from pvg.scenario_base.rollout_analysis import (
+    PureTextRolloutAnalyser,
+    ROLLOUT_ANALYSERS,
+)
 from pvg.utils.nested_array_dict import NestedArrayDict
-
-# TODO: Abstract this
-from pvg.code_validation.rollout_analysis import RolloutAnalyser, ROLLOUT_ANALYSERS
 
 
 @register_trainer(TrainerType.PURE_TEXT_EI)
@@ -178,10 +178,12 @@ class PureTextEiTrainer(PureTextRlTrainer):
 
     def run_analysers(
         self,
-        analysers: list[str | type[RolloutAnalyser]],
+        analysers: list[str | type[PureTextRolloutAnalyser]],
         model_name: str,
+        *,
         overwrite=False,
         use_tqdm=True,
+        dry_run=False,
     ):
         """Run the given analysers on the rollouts of the experiment.
 
@@ -189,22 +191,26 @@ class PureTextEiTrainer(PureTextRlTrainer):
 
         Parameters
         ----------
-        analysers : list[str | type[RolloutAnalyser]]
+        analysers : list[str | type[PureTextRolloutAnalyser]]
             The analysers to run. Either the name of the analyser or the analyser class
             itself.
         model_name : str
             The name of the model to use for the analysis.
-        overwrite : bool, optional
+        overwrite : bool, default=False
             Whether to overwrite the existing analysis files, if they exist.
-        use_tqdm : bool, optional
+        use_tqdm : bool, default=True
             Whether create a progress bar for the analysis.
+        dry_run : bool, default=False
+            Whether to do a dry run using a dummy API, not saving the results.
         """
 
         for analyser_cls in analysers:
 
             if isinstance(analyser_cls, str):
                 try:
-                    analyser_cls = ROLLOUT_ANALYSERS[analyser_cls]
+                    analyser_cls: type[PureTextRolloutAnalyser] = ROLLOUT_ANALYSERS[
+                        self.params.scenario, analyser_cls
+                    ]
                 except KeyError:
                     raise ValueError(
                         f"Analyser {analyser_cls!r} not found in list of analysers."
@@ -215,6 +221,7 @@ class PureTextEiTrainer(PureTextRlTrainer):
                 settings=self.settings,
                 protocol_handler=self.scenario_instance.protocol_handler,
                 model_name=model_name,
+                use_dummy_api=dry_run,
             )
 
             analysis_dir = self.checkpoint_analysis_dir.joinpath(analyser_cls.name)
@@ -239,7 +246,8 @@ class PureTextEiTrainer(PureTextRlTrainer):
                         self.settings.logger.warning(
                             f"Overwriting existing analysis file {analysis_file!r}"
                         )
-                    analysis_file.unlink()
+                    if not dry_run:
+                        analysis_file.unlink()
 
                 try:
                     rollouts = self.load_rollouts(iteration)
@@ -251,8 +259,9 @@ class PureTextEiTrainer(PureTextRlTrainer):
 
                 evaluations = analyser.forward(rollouts, use_tqdm=use_tqdm)
 
-                with open(analysis_file, "wb") as f:
-                    pickle.dump(evaluations, f)
+                if not dry_run:
+                    with open(analysis_file, "wb") as f:
+                        pickle.dump(evaluations, f)
 
     def _get_log_stats(
         self,
